refactor: _RagMarkdownConverter を共有モジュールに切り出す (#179)
